doc/requirements.txt
pyyaml
sphinx
tabulate
